Change API a little + fix netbsd timer
